Extension for constraints created, code reduce and simplify
It needs to verify the rotation of preview, it is not working with a
180 degrees rotation.
